poller/requirements.txt
feedparser==6.0.11
requests==2.31.0
APScheduler==3.10.4
nltk==3.8.1
python-dotenv==1.0.0
vaderSentiment==3.3.2
psycopg2-binary
